refactor(x): move xchat sources under src/chat
