+++ packages/adapter-x/src/chat/cards.ts
