+++ packages/adapter-x/src/chat/markdown.ts
