+++ packages/adapter-x/src/chat/types.ts
