+++ packages/adapter-x/tests/fixtures/sdk_vectors.json
