--- packages/adapter-xchat/vitest.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: "node",
-    coverage: {
-      provider: "v8",
-      reporter: ["text", "json-summary"],
-      include: ["src/**/*.ts"],
-      exclude: ["src/**/*.test.ts"],
-    },
-  },
-});
